requirements.txt
opencv-python
langchain
langchain-openai
langchain_google_genai
langchain-community
python-dotenv
pyaudio
soundfile
SpeechRecognition
SpeechRecognition[whisper-local]
SpeechRecognition[faster-whisper]
faster-whisper
mss
Pillow
Whisper
